pyproject.toml
[project]
name = "toy-issues-with-github-actions"
version = "0.1.0"
description = """
                A toy project to demonstrate automatic resolution of LLMOps issues with GitHub Actions.
            """
requires-python = ">=3.9,<3.10"
readme = "README.md"
authors = [
    { name = "Bruno Scaglione", email = "bruno.c.scaglione@gmail.com" }
]
license = {file = "LICENSE"}
dependencies = [
    "bcrypt>=4.2.1",
    "fastapi>=0.115.6",
    "llmops-issue-resolver>=0.1.1",
    "passlib>=1.7.4",
    "pydentic>=0.0.1.dev3",
    "python-jose>=3.3.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.34.0",
]

[[tool.uv.index]]
url = "https://test.pypi.org/simple/"
default = true

[[tool.uv.index]]
url = "https://pypi.org/simple/"

